Changes to be committed: new file:   CONTRIBUTING.md (Added how to add new source)
modified:   README.md (Link to CONTRIBUTING.md)

modified:   manga_dl/app.py (fix some bugs) [Todo: Fix  the slider]

modified:   manga_dl/main.py (nothing)

modified:   manga_dl/manga.py (fix chapter_exists, removed unnecessary imports, added clear_cache method)

modified:   manga_dl/manga_sources/__init__.py (nothing)

modified:   manga_dl/manga_sources/base_source.py (now id_to_url will be based on manga_format)

modified:   manga_dl/manga_sources/source1.py (uppercas R in refferer)

new file:   manga_dl/manga_sources/source2.py (Failed to add source2)

modified:   manga_dl/manga_sources/utils.py (equl id tweek)

modified:   manga_dl/public/manga.js (fixed time for 'a days ago')

modified:   manga_dl/templates/chapter.html (get to post for /api/img_url)

modified:   manga_dl/templates/manga.html (descending order for chapters)

modified:   manga_dl/tools/downloader2.py (check url is file or not)

modified:   manga_dl/tools/utils.py (adding option in driver)

modified:   requirements.txt (added Brotli>=1.0.9)

modified:   setup.py (version)

+++ manga_dl/tools/downloader2.py
@@ -1,4 +1,6 @@
 import os
+import shutil
+import re
 import json
 import logging
 from multiprocessing import Manager
@@ -41,7 +43,7 @@ class Downloader:
     ):
         self.urls = urls
         self.headers = headers or Headers().generate()
-        self.download_dir = download_dir or os.path.join(os.getcwd(), "downloads")
+        self.download_dir = download_dir or os.path.join(os.getcwd(), "tmp")
         if not os.path.exists(self.download_dir):
             os.mkdir(self.download_dir)
 
@@ -50,20 +52,37 @@ class Downloader:
         self.current_progress = 0
         self.total_urls = len(urls)
 
-    def get_filename(self, url):
-        return hashlib.sha1(url.encode()).hexdigest() + ".jpg"
+    @staticmethod
+    def is_file(url):
+        # check if url is not a url
+        if not re.match(r"^https?://", url):
+            if os.path.exists(url) or os.path.exists(compress_file_path(url)):
+                return True
+        return False
 
     async def download_file(self, session: aiohttp.ClientSession, url: str, pbar):
-        filename = self.get_filename(url)
-        filepath = os.path.join(self.download_dir, filename)
+        url = url.strip()
+        if not self.is_file(url):
+            filename = get_file_name(url)
+            filepath = os.path.join(self.download_dir, filename)
+        else:
+            filepath = url
+
+        # print("Downloading", url, filepath)
+
         cmp_filepath = compress_file_path(filepath)
+        isCompressed = False
+        isFileExists = False
         if os.path.exists(filepath):
-            self.downloaded_files.append(URLFile(url, filepath))
+            isFileExists = True
 
         elif os.path.exists(cmp_filepath):
-            self.downloaded_files.append(URLFile(url, cmp_filepath))
+            isFileExists = True
+            filepath = cmp_filepath
+            isCompressed = True
 
-        else:
+        failed = False
+        if not isFileExists:
             tmp_path = filepath + ".tmp"
             try:
                 timeout = aiohttp.ClientTimeout(
@@ -74,21 +93,23 @@ class Downloader:
                         async for chunk in response.content.iter_chunked(1024 * 1024):
                             if chunk:
                                 await f.write(chunk)
-
-                os.rename(tmp_path, filepath)
-                compressed = jpeg_compress(filepath, cmp_filepath)
-                if compressed:
-                    os.remove(filepath)
-                    filepath = cmp_filepath
-                else:
-                    raise Exception(f"Failed to compress {filepath}")
-
-                self.downloaded_files.append(URLFile(url, filepath))
-
+                shutil.move(tmp_path, filepath)
             except Exception as e:
-                print(f"Failed to download {url}: {e}")
+                logging.error(f"Failed to download {url}: {e}")
                 self.failed_urls.append(url)
-                return
+                failed = True
+
+        if not isCompressed and not failed:
+            compressed = jpeg_compress(filepath, cmp_filepath)
+            if compressed:
+                os.remove(filepath)
+                filepath = cmp_filepath
+            else:
+                logger.error(f"Failed to compress {filepath} {url}")
+
+        if not failed:
+            self.downloaded_files.append(URLFile(url, filepath))
+
         pbar.update(1)
         share_progress_bar(pbar.total, pbar.n, pbar.desc)
         self.total_urls -= 1
@@ -106,34 +127,54 @@ class Downloader:
                 await asyncio.gather(*tasks)
 
     @staticmethod
-    def download_one(url, headers, download_dir):
-        filename = get_file_name(url)
-        filepath = os.path.join(download_dir, filename)
-        cmp_filepath = compress_file_path(filepath)
-        if os.path.exists(filepath):
-            return URLFile(url, filepath)
-        elif os.path.exists(cmp_filepath):
-            return URLFile(url, cmp_filepath)
-
+    def download_one(url, headers, download_dir) -> URLFile:
+        url = url.strip()
+        if not Downloader.is_file(url):
+            filename = get_file_name(url)
+            filepath = os.path.join(download_dir, filename)
         else:
+            filepath = url
+
+        cmp_filepath = compress_file_path(filepath)
+        isCompressed = False
+        isFileExists = False
+        if os.path.exists(filepath):
+            isFileExists = True
+
+        elif os.path.exists(cmp_filepath):
+            isFileExists = True
+            filepath = cmp_filepath
+            isCompressed = True
+
+        failed = False
+        if not isFileExists:
             try:
                 response = requests.get(url, headers=headers)
                 with open(filepath, "wb") as f:
                     for chunk in response.iter_content(chunk_size=1024 * 1024):
                         if chunk:
                             f.write(chunk)
-                compressed = jpeg_compress(filepath, cmp_filepath)
-                if compressed:
-                    os.remove(filepath)
-                    filepath = cmp_filepath
-                return URLFile(url, filepath)
+
             except Exception as e:
                 logging.error(f"Failed to download {url}: {e}")
-                filepath = os.path.join(download_dir, get_file_name(f"{url}_failed"))
-                if not os.path.exists(filepath):
-                    create_failure_image(filepath, url)
+                failed = True
 
-                return URLFile(url, filepath)
+        if not isCompressed and not failed:
+            compressed = jpeg_compress(filepath, cmp_filepath)
+            if compressed:
+                os.remove(filepath)
+                filepath = cmp_filepath
+            else:
+                logger.error(f"Failed to compress {filepath} {url}")
+
+        if not failed:
+            return URLFile(url, filepath)
+        else:
+            filepath = os.path.join(download_dir, get_file_name(f"{url}_failed"))
+            if not os.path.exists(filepath):
+                create_failure_image(filepath, url)
+
+            return URLFile(url, filepath)
 
     def delete_tmp_files(self):
         for file in os.listdir(self.download_dir):
